fix various issues from commit 3509530
ref:db5ceb703ee5d9f04ebc27fd1be5d0bd1e3bbd0c

+++ catboost/libs/algo/split.h
@@ -164,7 +164,7 @@ struct TSplitTree {
 };
 
 struct TTreeStats {
-    TVector<float> LeafWeightsSum;
+    TVector<double> LeafWeightsSum;
 
     Y_SAVELOAD_DEFINE(LeafWeightsSum);
 };
